TINY-7829: moved align commands to a separate module
